requirements.txt
requests==2.30
python-dotenv==1.0.0
schedule==1.2.0